requirements.txt
fastapi~=0.115.4
paho-mqtt~=2.1.0
pydantic-settings~=2.6.1
uvicorn[standard]~=0.32.0